[CourseStatistics] Rewrite course search with Material UI
