--- services/frontend/src/pages/CourseStatistics/CourseTable/index.jsx
@@ -1,68 +1,0 @@
-import { sortBy } from 'lodash'
-import { arrayOf, bool, func, shape, string } from 'prop-types'
-import { memo } from 'react'
-import { Table } from 'semantic-ui-react'
-
-import { useLanguage } from '@/components/LanguagePicker/useLanguage'
-import { getActiveYears } from '../courseStatisticsUtils'
-import './courseTable.css'
-
-const CourseTable = ({ courses, onSelectCourse, hidden, title, controlIcon }) => {
-  const { getTextIn } = useLanguage()
-  const noContent = courses.length === 0
-  const sortCourses = courses => sortBy(courses, course => getTextIn(course.name))
-
-  const getEmptyListRow = () => (
-    <Table.Row>
-      <Table.Cell colSpan="3" content="No results." />
-    </Table.Row>
-  )
-
-  const toCourseRow = course => (
-    <Table.Row
-      key={course.id}
-      onClick={() => (course.min_attainment_date ? onSelectCourse(course) : null)}
-      style={{ cursor: 'pointer' }}
-    >
-      <Table.Cell width={10}>
-        <div>{getTextIn(course.name)}</div>
-        <div>{getActiveYears(course)}</div>
-      </Table.Cell>
-      <Table.Cell content={[course.code, ...course.substitutions].join(', ')} />
-      {controlIcon ? <Table.Cell icon={controlIcon} /> : null}
-    </Table.Row>
-  )
-
-  return (
-    !hidden && (
-      <Table className="fixed-header" selectable>
-        <Table.Header>
-          <Table.Row>
-            <Table.HeaderCell>{title}</Table.HeaderCell>
-            <Table.HeaderCell>Code</Table.HeaderCell>
-            {controlIcon ? <Table.HeaderCell /> : null}
-          </Table.Row>
-        </Table.Header>
-        <Table.Body>{noContent ? getEmptyListRow() : sortCourses(courses).map(toCourseRow)}</Table.Body>
-      </Table>
-    )
-  )
-}
-
-CourseTable.propTypes = {
-  courses: arrayOf(shape({ code: string, name: shape({}), seleted: bool })).isRequired,
-  onSelectCourse: func.isRequired,
-  hidden: bool.isRequired,
-  title: string.isRequired,
-
-  controlIcon: string,
-}
-
-const areEqual = (prevProps, nextProps) => {
-  if (prevProps.courses.length !== nextProps.courses.length) {
-    return false
-  }
-  return prevProps.courses.every(c1 => nextProps.courses.some(c2 => c1.code === c2.code))
-}
-
-export const MemoizedCourseTable = memo(CourseTable, areEqual)
